Fix / change to use ChatGPT for text completion
Current model/call returned an error:

openai.error.InvalidRequestError: This is a chat model and not
supported in the v1/completions endpoint. Did you mean to use
v1/chat/completions?

There was a mix between using the chat and the text api.
I didn't know which was intended, so I fixed it using the
ChatCompletion one (a.k.a. ChatGPT 3 api).

There is a better way to use this (combining system and user
roles) but that will be changed on a followup commit.