Implement db-based tracking to handle tweets
Manage which twits were sent and which are pending thanks to a
small sqlite3 db file.

The bot now goes to sleep and every 300 seconds it checks if
there are new sources and send 1 twit about it, then goes back
to sleep.

There are 2 implemented sources right now:

* NBA scoreboards
* Marca Kings League-related news

In the future the db can be used to store the persistent
configuration as well as more dynamic prompts.