Fix error handling when a recap is invalid
The last Celtics recap is missing. Fix that edge case where a
source returns invalid input data, by tracking the errors, and
retrying it only 3 times. Sources that have failed 3 or more times
are ignored.

This requires changing the sources table schema to include a new
column "errors" that tracks how many errors a particular source
has returned, so it can be tracked over time and persisted to
disk.